+++ .idea/assetWizardSettings.xml
@@ -28,10 +28,9 @@
                         </option>
                         <option name="values">
                           <map>
-                            <entry key="outputName" value="ic_baseline_star_border_24" />
                             <entry key="assetSourceType" value="FILE" />
-                            <entry key="outputName" value="naver_login_btn" />
-                            <entry key="sourceFile" value="C:\Users\korea\Desktop\naver_login_btn.svg" />
+                            <entry key="outputName" value="ic_setting_line" />
+                            <entry key="sourceFile" value="$USER_HOME$/Downloads/Setting_line.svg" />
                           </map>
                         </option>
                       </PersistentState>
