--- app/src/main/java/com/example/ziririt/presentation/board/streamer/PostItem.kt
@@ -1,9 +1,0 @@
-package com.example.ziririt.presentation.board.streamer
-
-data class PostItem(
-    val image: Int,
-    val title: String,
-    val commentNum: Int,
-    val name: String,
-    val date: String,
-)
