+++ app/src/main/java/com/example/ziririt/presentation/board/streamer_board/PostItem.kt
@@ -1,0 +1,9 @@
+package com.example.ziririt.presentation.board.streamer_board
+
+data class PostItem(
+    val image: Int,
+    val title: String,
+    val commentNum: Int,
+    val name: String,
+    val date: String,
+)
